Move storage tests to tests folder
