+++ src/redisearch_rs/reducers/src/collect/mod.rs
@@ -23,7 +23,8 @@
 pub(crate) mod common;
 pub mod local;
 pub mod remote;
-pub(crate) mod storage;
+#[doc(hidden)]
+pub mod storage;
 
 pub use local::{LocalCollectCtx, LocalCollectReducer};
 pub use remote::{RemoteCollectCtx, RemoteCollectReducer};
